Move `FusionJarInfo` into `runtime.base`.
I'm not sure that this class is really needed, and I'm not sure that this is where it should be, but I'm sure that it shouldn't be under `dev.ionfusion.fusion`.
