+++ fusioncli/src/main/java/dev/ionfusion/fusioncli/Version.java
@@ -8,8 +8,8 @@ import com.amazon.ion.IonType;
 import com.amazon.ion.IonWriter;
 import com.amazon.ion.system.IonTextWriterBuilder;
 import com.amazon.ion.util.JarInfo;
-import dev.ionfusion.fusion.util.FusionJarInfo;
 import dev.ionfusion.runtime.base.FusionException;
+import dev.ionfusion.runtime.base.FusionJarInfo;
 import java.io.IOException;
 import java.io.PrintWriter;
